Fixing an issue with renderjobjessica not returning the width and height in an int format, therefore breaking asset image generation https://jira.ccpgames.com/browse/EO-35660
[git-p4: depot-paths = "//eve/branches/development/MAINLINE/packages/trinity/": change = 2137739]

+++ sceneRenderJobSpaceJessica.py
@@ -118,7 +118,7 @@ class SceneRenderJobSpaceJessica(SceneRenderJobSpace):
         
         width = self.backBufferOverride.width
         height = self.backBufferOverride.height
-        return width * self.invBackBufferScale, height * self.invBackBufferScale
+        return int(width * self.invBackBufferScale), int(height * self.invBackBufferScale)
 
     def GetBackBufferRenderTarget(self):
         """
